ai/requirements_simple.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
